feat: add participant update and delete

+++ src/routes/[id]/components/DeleteExpenseDialog.svelte
@@ -1,27 +1,29 @@
 <script lang="ts">
   import { page } from '$app/stores';
   import ConfirmationDialog from '$lib/components/ConfirmationDialog.svelte';
-  import type createBill from '../stores/bill.svelte';
+  import type createBill from '../stores/bill';
   import { getContext } from 'svelte';
 
-  let { onclose }: { onclose: () => void } = $props();
+  let { onclose, open }: { onclose: () => void; open: boolean } = $props();
 
-  const { expenses } = getContext<ReturnType<typeof createBill>>('bill');
+  const { bill, deleteExpense } = getContext<ReturnType<typeof createBill>>('bill');
   let loading = $state(false);
 
-  let { expenseId = '' } = $page.state;
+  let { expenseId } = $page.state;
 
   const handleDelete = async () => {
+    if (expenseId === undefined) return;
+
     await fetch(`/api/bills/${$page.params.id}/expenses/${expenseId}`, { method: 'DELETE' });
-    delete expenses[expenseId];
+    deleteExpense(expenseId);
     onclose();
   };
 </script>
 
 {#if expenseId}
-  {@const expense = expenses[expenseId]}
+  {@const expense = $bill.expenses.find(({ id }) => id === expenseId)}
   <ConfirmationDialog
-    open={$page.state.modalShown === 'expense'}
+    {open}
     title="Delete bill"
     message={`Are you sure you want to delete the expense ${expense?.title}?`}
     disabled={loading}
